+++ app/src/main/java/woowacourse/shopping/feature/productlist/ProductListUiState.kt
@@ -1,0 +1,12 @@
+package woowacourse.shopping.feature.productlist
+
+import woowacourse.shopping.feature.common.state.ProductUiModel
+
+data class ProductListUiState(
+    val productUiModels: List<ProductUiModel> = emptyList(),
+    val recentProducts: List<ProductUiModel> = emptyList(),
+    val mostRecentProductId: Long? = null,
+    val isLoading: Boolean = true,
+    val isEnd: Boolean = false,
+    val cartTotalQuantity: Int = 0,
+)
